requirements.txt
# ── Orchestration ─────────────────────────────────────────────────────────────
dagster==1.9.4
dagster-webserver==1.9.4
dagster-dbt==0.25.4

# ── Transformation (dbt) ──────────────────────────────────────────────────────
dbt-core==1.7.9
dbt-mysql==1.7.0

# ── Database drivers ──────────────────────────────────────────────────────────
sqlalchemy==2.0.36
mysqlclient==2.2.4
PyMySQL==1.1.1

# ── Data ingestion ────────────────────────────────────────────────────────────
sling==1.5.13

# ── Dummy data generator ──────────────────────────────────────────────────────
faker==33.1.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pyyaml==6.0.2
protobuf>=4.25.0,<6.0.0
grpcio>=1.62.0
grpcio-health-checking>=1.62.0
